Fix boot sequence and periodic sync reliability issues
- Re-enable brownout detector after BLE RF calibration completes
- Fix BLE stop to properly wait for task termination with timeout
- Use 64-bit esp_timer timestamps to prevent wrap after ~49 days
- Start BLE task before timers to avoid initialization race
- Remove duplicate NVS initialization from inkbird_ble_init
- Protect s_peer_count modifications with mutex
- Reorder state init before semaphore drain to close race window
- Clear task handle before vTaskDelete for clean shutdown signaling

+++ src/ble/inkbird_ble.h
@@ -37,7 +37,7 @@ typedef struct {
     int16_t  temperature;   // Temperature in 0.1°C units
     uint16_t humidity;      // Relative humidity in 0.1% units
     uint16_t pressure;      // Atmospheric pressure in hPa
-    uint32_t timestamp;     // Reading timestamp (ms since boot)
+    int64_t  timestamp;     // Reading timestamp (ms since boot, from esp_timer)
     bool     valid;         // true if data has been received at least once
     bool     stale;         // true if data is older than INKBIRD_DATA_STALE_MS
 } inkbird_reading_t;
